Ordinal scale method definition

A general function to define a method for measurements in an ordinal scale of a given subject.

defineOrdinalScaleMethod(name, description, subject, codes,
  citationString = "", DOI = "", quantifiableCodes = character(0),
  breaks = NULL, midPoints = NULL, definitions = NULL)

Arguments

name
String with the name of the method.

description
String describing the method.

subject
A string to identify the subject (e.g. 'plant cover').

codes
A character vector of ordinal class codes.

citationString
A string with the bibliographic reference for the method.

DOI
A string with the DOI the resource related to citationString.

quantifiableCodes
A string vector with the subset of codes that have correspondence in a quantitative scale. These codes can be transformed to midPoints (if supplied in midPoints) and, correspondingly, codes in the quantitative scale can be transformed to ordinal codes using class limits (if supplied in breaks).

breaks
A vector of break points (for ordinal scales with class limits)

midPoints
A vector of class midpoints (for ordinal scales that can be translated into quantitative values)

definitions
A character vector of class definitions.

Value

an object of class VegXMethodDefinition

See also

Examples

#Ordinal scale with three levels
defineOrdinalScaleMethod("scale1", "Description for scale1", "subject1",
                   codes = c("first", "second","third"))

#> An object of class "VegXMethodDefinition"
#> Slot "name":
#> [1] "scale1"
#> 
#> Slot "description":
#> [1] "Description for scale1"
#> 
#> Slot "citationString":
#> [1] ""
#> 
#> Slot "DOI":
#> [1] ""
#> 
#> Slot "subject":
#> [1] "subject1"
#> 
#> Slot "attributeType":
#> [1] "ordinal"
#> 
#> Slot "attributes":
#> $`1`
#> $`1`$type
#> [1] "ordinal"
#> 
#> $`1`$code
#> [1] "first"
#> 
#> $`1`$order
#> [1] 1
#> 
#> 
#> $`2`
#> $`2`$type
#> [1] "ordinal"
#> 
#> $`2`$code
#> [1] "second"
#> 
#> $`2`$order
#> [1] 2
#> 
#> 
#> $`3`
#> $`3`$type
#> [1] "ordinal"
#> 
#> $`3`$code
#> [1] "third"
#> 
#> $`3`$order
#> [1] 3
#> 
#> 
#> 

#Braun-Blanquet plant cover scale with five levels that can be translated to cover codes
defineOrdinalScaleMethod(name = "Braun-Blanquet",
                   description = "Five-level Braun-Blanquet cover scale",
                   subject = "plant cover",
                   codes = as.character(1:5),
                   quantifiableCodes = as.character(1:5),
                   breaks = c(0,5,25,50,75,100),
                   midPoints = c(2.5,17.5, 37.5, 62.5, 87.5))

#> An object of class "VegXMethodDefinition"
#> Slot "name":
#> [1] "Braun-Blanquet"
#> 
#> Slot "description":
#> [1] "Five-level Braun-Blanquet cover scale"
#> 
#> Slot "citationString":
#> [1] ""
#> 
#> Slot "DOI":
#> [1] ""
#> 
#> Slot "subject":
#> [1] "plant cover"
#> 
#> Slot "attributeType":
#> [1] "ordinal"
#> 
#> Slot "attributes":
#> $`1`
#> $`1`$type
#> [1] "ordinal"
#> 
#> $`1`$code
#> [1] "1"
#> 
#> $`1`$order
#> [1] 1
#> 
#> $`1`$lowerLimit
#> [1] 0
#> 
#> $`1`$upperLimit
#> [1] 5
#> 
#> $`1`$midPoint
#> [1] 2.5
#> 
#> 
#> $`2`
#> $`2`$type
#> [1] "ordinal"
#> 
#> $`2`$code
#> [1] "2"
#> 
#> $`2`$order
#> [1] 2
#> 
#> $`2`$lowerLimit
#> [1] 5
#> 
#> $`2`$upperLimit
#> [1] 25
#> 
#> $`2`$midPoint
#> [1] 17.5
#> 
#> 
#> $`3`
#> $`3`$type
#> [1] "ordinal"
#> 
#> $`3`$code
#> [1] "3"
#> 
#> $`3`$order
#> [1] 3
#> 
#> $`3`$lowerLimit
#> [1] 25
#> 
#> $`3`$upperLimit
#> [1] 50
#> 
#> $`3`$midPoint
#> [1] 37.5
#> 
#> 
#> $`4`
#> $`4`$type
#> [1] "ordinal"
#> 
#> $`4`$code
#> [1] "4"
#> 
#> $`4`$order
#> [1] 4
#> 
#> $`4`$lowerLimit
#> [1] 50
#> 
#> $`4`$upperLimit
#> [1] 75
#> 
#> $`4`$midPoint
#> [1] 62.5
#> 
#> 
#> $`5`
#> $`5`$type
#> [1] "ordinal"
#> 
#> $`5`$code
#> [1] "5"
#> 
#> $`5`$order
#> [1] 5
#> 
#> $`5`$lowerLimit
#> [1] 75
#> 
#> $`5`$upperLimit
#> [1] 100
#> 
#> $`5`$midPoint
#> [1] 87.5
#> 
#> 
#> 

#Ordinal plant cover scale with seven levels where six can be translated to a quantitative scale
defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007) 
                        The Recce method for describing New Zealand vegetation – Field protocols. 
                        Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1%", "1-5%","6-25%", "26-50%", 
                                   "51-75%", "76-100%"))

#> An object of class "VegXMethodDefinition"
#> Slot "name":
#> [1] "Recce cover scale"
#> 
#> Slot "description":
#> [1] "Recce recording method by Hurst/Allen"
#> 
#> Slot "citationString":
#> [1] "Hurst, JM and Allen, RB. (2007) \n                        The Recce method for describing New Zealand vegetation – Field protocols. \n                        Landcare Research, Lincoln."
#> 
#> Slot "DOI":
#> [1] ""
#> 
#> Slot "subject":
#> [1] "plant cover"
#> 
#> Slot "attributeType":
#> [1] "ordinal"
#> 
#> Slot "attributes":
#> $`1`
#> $`1`$type
#> [1] "ordinal"
#> 
#> $`1`$code
#> [1] "P"
#> 
#> $`1`$order
#> [1] 1
#> 
#> $`1`$definition
#> [1] "Presence"
#> 
#> 
#> $`2`
#> $`2`$type
#> [1] "ordinal"
#> 
#> $`2`$code
#> [1] "1"
#> 
#> $`2`$order
#> [1] 2
#> 
#> $`2`$definition
#> [1] "<1%"
#> 
#> $`2`$lowerLimit
#> [1] 0
#> 
#> $`2`$upperLimit
#> [1] 1
#> 
#> $`2`$midPoint
#> [1] 0.05
#> 
#> 
#> $`3`
#> $`3`$type
#> [1] "ordinal"
#> 
#> $`3`$code
#> [1] "2"
#> 
#> $`3`$order
#> [1] 3
#> 
#> $`3`$definition
#> [1] "1-5%"
#> 
#> $`3`$lowerLimit
#> [1] 1
#> 
#> $`3`$upperLimit
#> [1] 5
#> 
#> $`3`$midPoint
#> [1] 0.5
#> 
#> 
#> $`4`
#> $`4`$type
#> [1] "ordinal"
#> 
#> $`4`$code
#> [1] "3"
#> 
#> $`4`$order
#> [1] 4
#> 
#> $`4`$definition
#> [1] "6-25%"
#> 
#> $`4`$lowerLimit
#> [1] 5
#> 
#> $`4`$upperLimit
#> [1] 25
#> 
#> $`4`$midPoint
#> [1] 15
#> 
#> 
#> $`5`
#> $`5`$type
#> [1] "ordinal"
#> 
#> $`5`$code
#> [1] "4"
#> 
#> $`5`$order
#> [1] 5
#> 
#> $`5`$definition
#> [1] "26-50%"
#> 
#> $`5`$lowerLimit
#> [1] 25
#> 
#> $`5`$upperLimit
#> [1] 50
#> 
#> $`5`$midPoint
#> [1] 37.5
#> 
#> 
#> $`6`
#> $`6`$type
#> [1] "ordinal"
#> 
#> $`6`$code
#> [1] "5"
#> 
#> $`6`$order
#> [1] 6
#> 
#> $`6`$definition
#> [1] "51-75%"
#> 
#> $`6`$lowerLimit
#> [1] 50
#> 
#> $`6`$upperLimit
#> [1] 75
#> 
#> $`6`$midPoint
#> [1] 62.5
#> 
#> 
#> $`7`
#> $`7`$type
#> [1] "ordinal"
#> 
#> $`7`$code
#> [1] "6"
#> 
#> $`7`$order
#> [1] 7
#> 
#> $`7`$definition
#> [1] "76-100%"
#> 
#> $`7`$lowerLimit
#> [1] 75
#> 
#> $`7`$upperLimit
#> [1] 100
#> 
#> $`7`$midPoint
#> [1] 87.5
#> 
#> 
#>