Quantitative scale method definition

A general function to define a method for measurements in a quantitative scale of a given subject

defineQuantitativeScaleMethod(
  name,
  description,
  subject,
  unit,
  citationString = "",
  DOI = "",
  lowerLimit = -Inf,
  upperLimit = Inf
)

Arguments

name

String with the name of the method.

description

String describing the method.

subject

A string to identify the subject.

unit

A character describing measurement units.

citationString

A string with the bibliographic reference for the method.

DOI

A string with the DOI the resource related to citationString

lowerLimit

The lower limit of the quantitative scale, if defined.

upperLimit

The upper limit of the quantitative scale, if defined.

Value

an object of class VegXMethodDefinition

See also

Examples

defineQuantitativeScaleMethod(name = "Percent cover",
                              description = "Quantitative plant percent cover",
                              unit = "%",
                              subject = "plant cover",
                              lowerLimit = 0, upperLimit = 100)
#> An object of class "VegXMethodDefinition"
#> Slot "name":
#> [1] "Percent cover"
#> 
#> Slot "description":
#> [1] "Quantitative plant percent cover"
#> 
#> Slot "citationString":
#> [1] ""
#> 
#> Slot "DOI":
#> [1] ""
#> 
#> Slot "subject":
#> [1] "plant cover"
#> 
#> Slot "attributeType":
#> [1] "quantitative"
#> 
#> Slot "attributes":
#> $`1`
#> $`1`$type
#> [1] "quantitative"
#> 
#> $`1`$unit
#> [1] "%"
#> 
#> $`1`$lowerLimit
#> [1] 0
#> 
#> $`1`$upperLimit
#> [1] 100
#> 
#> 
#>