Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 5b82b84e-5522-4ed6-b630-507e46380c3d            <NA>
#> 2  LGM08r_1Q 72a8f51b-0ba0-421c-b7f4-1ae0c5e328f7          LGM08r
#> 3  LGM08r_2Q 5f9c807b-9f32-4a0d-ae56-a1f0dd6e0b83          LGM08r
#> 4  LGM08r_3Q eb1e4415-d8d7-40f7-a439-a3424f90583e          LGM08r
#> 5  LGM08r_4Q 863b213a-6547-437b-ad87-f1ca2e5d98f2          LGM08r
#> 6     LGM16l 3ea877a9-8df0-4b5a-bb13-4b2881fa7a91            <NA>
#> 7  LGM16l_1Q 14e4170a-e80d-47a6-93b8-416ef84ec861          LGM16l
#> 8  LGM16l_2Q cab0acdd-999c-4c76-937a-741f99edd12e          LGM16l
#> 9  LGM16l_3Q be07f96b-b7fb-42c5-b9c7-a44b4203d63b          LGM16l
#> 10 LGM16l_4Q 4f5d2550-5a9a-4768-aadd-fee941e214e4          LGM16l
#> 11    LGM38h 4cc7df2e-206b-4287-8443-d861a5c9a091            <NA>
#> 12 LGM38h_1Q e6bf2924-de55-4230-8acd-e021db71bd17          LGM38h
#> 13 LGM38h_2Q 5e0d1f3f-e90e-45dd-9735-5843a2750ff4          LGM38h
#> 14 LGM38h_3Q 90e8e7ce-5975-4ae4-9317-a3ac6ca82cf5          LGM38h
#> 15 LGM38h_4Q 2c01238d-8e1f-4306-82c9-bbe203d7410e          LGM38h
#> 16    LGM45h 65f939c6-0e13-471e-a2b7-3de1da8328a0            <NA>
#> 17 LGM45h_2Q 071bf148-1c49-4136-9bfb-be8c7603316f          LGM45h
#> 18 LGM45h_3Q 4abab55c-d106-4237-bf92-d9a39a23e07c          LGM45h
#> 19 LGM45h_4Q 702df258-9558-4304-aefb-7212bb8972ac          LGM45h
#> 20    LGM43h e6612b49-b2b4-4efd-9e33-7efe59cd5254            <NA>
#> 21 LGM43h_1Q 949e4539-b916-466d-80b2-9432894d8722          LGM43h
#> 22 LGM43h_2Q 7003e83a-f05d-49e2-b7d1-a168bde4a40b          LGM43h
#> 23 LGM43h_3Q 5db5c825-8ef3-4994-ada4-825abf0ccfd6          LGM43h
#> 24 LGM43h_4Q 38b1fa7c-bc1b-4a31-8950-9f5ca0864977          LGM43h
#> 25 LGM45h_1Q 83d48bc9-8ba8-4092-9ac7-7d05bd9453ec          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot