Predefined measurement methods

Creates a measurement method for Veg-X documents from a set of predefined options

predefinedMeasurementMethod(name)

Arguments

name
A string with the desired measurement method. Current accepted options are:

"Plot area/m2": Plot area measured in squared meters.

"Plot area/cm2": Plot area measured in squared centimeters.

"Plot dimension/m": Plot dimension (radius, length, width) measured in meters.

"Plot dimension/cm": Plot dimension (radius, length, width) measured in centimeters.

"Plant cover/%": Plant cover as percentage of ground covered by the projection.

"Plant counts": Number of plant individuals.

"Plant frequency/%": Frequency of occurrence in fixed subunits of the plot.

"Basal area/m2·ha-1": Basal area in square meters per hectare.

"Stratum height/m": Stratum height in meters above the ground.

"Stratum height/cm": Stratum height in cm above the ground.

"Plant height/m": Plant height in meters above the ground.

"Plant height/cm": Plant height in cm above the ground.

"DBH/cm": Diameter at breast height, in cm.

"Surface cover/%": Surface covered by a surface type as percentage of ground covered by the projection.

"pH": pH scale from 0 to 14.

"Slope/degrees": Slope measured using degrees.

"Aspect/degrees": Aspect measured using degrees from North.

"Elevation/m": Elevation measured using meters above sea level.

Value

an object of class VegXMethodDefinition

See also

Other define measurement functions: defineOrdinalScaleMethod, defineQualitativeScaleMethod, defineQuantitativeScaleMethod

Examples

# Create a method for plant percent cover
predefinedMeasurementMethod("Plant cover/%")

#> An object of class "VegXMethodDefinition"
#> Slot "name":
#> [1] "Plant cover/%"
#> 
#> Slot "description":
#> [1] "Plant cover as percentage of ground covered by the projection."
#> 
#> Slot "citationString":
#> [1] ""
#> 
#> Slot "DOI":
#> [1] ""
#> 
#> Slot "subject":
#> [1] "plant cover"
#> 
#> Slot "attributeType":
#> [1] "quantitative"
#> 
#> Slot "attributes":
#> $`1`
#> $`1`$type
#> [1] "quantitative"
#> 
#> $`1`$unit
#> [1] "%"
#> 
#> $`1`$lowerLimit
#> [1] 0
#> 
#> $`1`$upperLimit
#> [1] 100
#> 
#> 
#>