Sets original identification taxon concepts

Sets (or resets) taxon concepts associated with the original identification of organisms

setOriginalIdentificationConcepts(
  target,
  citationStringAll = "",
  x = NULL,
  mapping = list(),
  date.format = "%Y-%m-%d",
  missing.values = c(NA, ""),
  verbose = TRUE
)

Arguments

target

The initial object of class VegX to be modified

citationStringAll

A string with the bibliographic citation to be applied to all organism identities of the VegX object (using the original organism names as taxon names), or to all original organism names listed in x.

x

A data frame where each row corresponds to one organism identity, given by a column that can be mapped to originalOrganismName.

mapping

A named list whose elements are strings that correspond to column names in x. Names of the list should be:

originalOrganismName - A string with the original name given by the author of the data set (required).

conceptName - A string with the taxon name forming the taxon concept, if different from originalOrganismName (optional).

conceptCitation - A string with the bibliographic citation forming the taxon concept (optional).

assertionDate - Date of taxon concept assertion (see date.format) (optional).

assertionParty - Name of the party that undertook taxon concept assertion (optional).

date.format

A character string specifying the input format of dates (see as.Date).

missing.values

A character vector of values that should be considered as missing data (see details).

verbose

A boolean flag to indicate console output of the concept identification process.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create new Veg-X document with aggregate organism observations
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName",
               stratumName = "Tier", cover = "Category")
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007)
                        The Recce method for describing New Zealand vegetation – Field protocols.
                        Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1%", "1-5%","6-25%", "26-50%",
                                  "51-75%", "76-100%"))
strataDef = defineMixedStrata(name = "Recce strata",
                   description = "Standard Recce stratum definition",
                   citation = "Hurst, JM and Allen, RB. (2007)
                      The Recce method for describing New Zealand vegetation – Field protocols.
                      Landcare Research, Lincoln.",
                   heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                   heightStrataNames = paste0("Tier ",1:6),
                   categoryStrataNames = "Tier 7",
                   categoryStrataDefinition = "Epiphytes")
x = addAggregateOrganismObservations(newVegX(), moki_tcv,
                   mapping = mapping,
                   methods = c(cover=coverscale),
                   stratumDefinition = strataDef)
#>  1 additional aggregate organism measurements found: Category.
#>  Measurement method 'Recce cover scale' added for 'cover'.
#>  Stratum definition method 'Recce strata' added.
#>  7 new stratum definitions added.
#>  5 plot(s) parsed, 5 new added.
#>  5 plot observation(s) parsed, 5 new added.
#>  148 organism names(s) parsed, 148 new added.
#>  148 organism identitie(s) parsed, 148 new added.
#>  33 stratum observation(s) parsed, 33 new added.
#>  582 record(s) parsed, 582 new aggregate organism observation(s) added.

# Inspect the original organism identities
head(showElementTable(x, "organismIdentity"))
#>             identityName   originalOrganismName taxon
#> 1  Dacrydium cupressinum  Dacrydium cupressinum  TRUE
#> 2    Weinmannia racemosa    Weinmannia racemosa  TRUE
#> 3       Myrsine salicina       Myrsine salicina  TRUE
#> 4        Cyathea smithii        Cyathea smithii  TRUE
#> 5 Prumnopitys ferruginea Prumnopitys ferruginea  TRUE
#> 6    Nothofagus truncata    Nothofagus truncata  TRUE

y = setOriginalIdentificationConcepts(x, citationStringAll="Allen 1998")
#>  0 taxon concept(s) reset. 148 new taxon concept(s) added.
#>  0 new literature citation(s) added.

head(showElementTable(y, "organismIdentity"))
#>             identityName   originalOrganismName taxon
#> 1  Dacrydium cupressinum  Dacrydium cupressinum  TRUE
#> 2    Weinmannia racemosa    Weinmannia racemosa  TRUE
#> 3       Myrsine salicina       Myrsine salicina  TRUE
#> 4        Cyathea smithii        Cyathea smithii  TRUE
#> 5 Prumnopitys ferruginea Prumnopitys ferruginea  TRUE
#> 6    Nothofagus truncata    Nothofagus truncata  TRUE
#>                             taxonConcept
#> 1  Dacrydium cupressinum sec. Allen 1998
#> 2    Weinmannia racemosa sec. Allen 1998
#> 3       Myrsine salicina sec. Allen 1998
#> 4        Cyathea smithii sec. Allen 1998
#> 5 Prumnopitys ferruginea sec. Allen 1998
#> 6    Nothofagus truncata sec. Allen 1998