S4 class for Veg-X documents

Implements the minimum subset of Veg-X elements needed for importing and combining vegetation data in this package. Only main elements of Veg-X are included here as lists (sub-elements are not specified in the class definition). Other main Veg-X elements ('taxonDetermination', 'communityConcept', 'communityDetermination') are not yet implemented.

Usage

# S4 method for VegX
summary(object, ...)

Arguments

object

Veg-X object to be summarized

...

Additional parameters (for compatibility)

Methods (by generic)

summary: Generates a summary of the Veg-X object (i.e., the number of elements of each type).

Slots

VegXVersion

A string to indicate the version of the Veg-X standard.

parties

A list of persons, organisations mentioned in the data set.

literatureCitations

A list of literature citations used in the data set.

methods

A list of measurement methods (e.g. cover scales) mentioned in the data set.

attributes

A list of attribute value descriptions used in the data set.

strata

A list of stratum definitions used in the data set.

surfaceTypes

A list of surface class definitions used in the data set.

organismNames

A list of organism names used in the data set.

taxonConcepts

A list of taxon concepts (i.e. taxon name + citation) used in the data set.

organismIdentities

A list of organism identities used in the data set.

projects

A list of research projects underpinning plot data collection.

plots

A list of vegetation plots where measurements were made.

individualOrganisms

A list of individual organisms found in vegetation plots.

plotObservations

A list of observations (i.e. sampling events) made on vegetation plots.

individualObservations

A list of observations made on individual organisms (e.g. diameter or height values).

aggregateObservations

A list of observations made on aggregate organisms (e.g. plant abundance by taxa or stratum).

stratumObservations

A list of observations made on strata (e.g. overall tree cover).

communityObservations

list of observations applying to the whole plant community (e.g. stand age or species richness)

siteObservations

A list of observations that apply to the site (e.g., abiotic measurements, land use, management regime, legal protection, ...).

surfaceCoverObservations

list of surface cover observations.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

showClass("VegX")
#> Class "VegX" [package "VegX"]
#> 
#> Slots:
#>                                                         
#> Name:               VegXVersion                  parties
#> Class:                character                     list
#>                                                         
#> Name:       literatureCitations                  methods
#> Class:                     list                     list
#>                                                         
#> Name:                attributes                   strata
#> Class:                     list                     list
#>                                                         
#> Name:              surfaceTypes            organismNames
#> Class:                     list                     list
#>                                                         
#> Name:             taxonConcepts       organismIdentities
#> Class:                     list                     list
#>                                                         
#> Name:                  projects                    plots
#> Class:                     list                     list
#>                                                         
#> Name:       individualOrganisms         plotObservations
#> Class:                     list                     list
#>                                                         
#> Name:    individualObservations    aggregateObservations
#> Class:                     list                     list
#>                                                         
#> Name:       stratumObservations    communityObservations
#> Class:                     list                     list
#>                                                         
#> Name:          siteObservations surfaceCoverObservations
#> Class:                     list                     list