Add site observation records

Adds site observation records to a VegX object from a data table where rows are plot observations.

addSiteObservations(target, x, plotObservationMapping,
  soilMeasurementMapping = list(), climateMeasurementMapping = list(),
  waterBodyMeasurementMapping = list(), soilMeasurementMethods = list(),
  climateMeasurementMethods = list(), waterBodyMeasurementMethods = list(),
  date.format = "%Y-%m-%d", missing.values = c(NA, ""), verbose = TRUE)

Arguments

target
The initial object of class VegX to be modified

x
A data frame where each row corresponds to one plot observation. Columns can be varied.

plotObservationMapping
A list with element names 'plotName', 'obsStartDate', used to specify the mapping of data columns (specified using strings for column names) onto these variables. Additional optional mappings are: 'subPlotName'.

soilMeasurementMapping
A list with element names equal to soil measurement subjects, used to specify the mapping of data columns (specified using strings for column names) onto these variables.

climateMeasurementMapping
A list with element names equal to climate measurement subjects, used to specify the mapping of data columns (specified using strings for column names) onto these variables.

waterBodyMeasurementMapping
A list with element names equal to water body measurement subjects, used to specify the mapping of data columns (specified using strings for column names) onto these variables.

soilMeasurementMethods
A named list of objects of class VegXMethodDefinition with the measurement method for each of the soil variables stated in soilMeasurementMapping. List names should be the same as soil subject measurement variables (e.g. list(pH = pHmeth) to specify the use of method 'pHmeth' for pH measurements).

climateMeasurementMethods
A named list of objects of class VegXMethodDefinition with the measurement method for each of the soil variables stated in soilMeasurementMapping. List names should be the same as climate subject measurement variables.

waterBodyMeasurementMethods
A named list of objects of class VegXMethodDefinition with the measurement method for each of the soil variables stated in soilMeasurementMapping. List names should be the same as water body subject measurement variables.

date.format
A character string specifying the input format of dates (see as.Date).

missing.values
A character vector of values that should be considered as missing observations/measurements.

verbose
A boolean flag to indicate console output of the data integration process.

Value

The modified object of class VegX.

Details

Missing value policy:

Missing 'plotName' or 'obsStartDate' values are interpreted as if the previous non-missing value has to be used to define plot observation.

Missing 'subPlotName' values are interpreted in that observation refers to the parent plotName.

Missing measurements are simply not added to the Veg-X document.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Other add functions: addAggregateOrganismObservations, addCommunityObservations, addIndividualOrganismObservations, addPlotGeometries, addPlotLocations, addPlotObservations, addSiteCharacteristics, addStratumObservations, addSurfaceCoverObservations, addTaxonBySiteData

Examples

# Load source data
data(mokihinui)

# Define mapping
mapping = list(plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate")
soilmapping = list(pH = "pH")

# Define pH method
pHMeth = predefinedMeasurementMethod("pH")

# Create new Veg-X document with site observations
x = addSiteObservations(newVegX(), moki_site,
                        plotObservationMapping = mapping,
                        soilMeasurementMapping = soilmapping,
                        soilMeasurementMethods = list(pH = pHMeth))

#>  Measurement method 'pH' added for 'pH'.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.
#>  25 record(s) parsed, 25 new site observation(s) added.

# Examine results
summary(x)

#> ================================================================
#>                     Veg-X object (ver 2.0.0)                   
#> ----------------------------------------------------------------
#> 
#>    Projects: 0
#> 
#>    Plots: 25  [Parent plots: 5 Sub-plots: 20]
#> 
#>    Individual organisms: 0
#> 
#>    Organism names: 0
#> 
#>    Taxon concepts: 0
#> 
#>    Organism Identities: 0
#> 
#>    Vegetation strata: 0
#> 
#>    Surface types: 0
#> 
#>    Parties: 0
#> 
#>    Literature citations: 0
#> 
#>    Methods: 1
#>       1. pH [pH / 1 quantitative atts.]
#> 
#>    Plot observations: 25  [in parent plots: 5 in sub-plots: 20]
#> 
#>    Individual organism observations: 0
#> 
#>    Aggregated organism observations: 0
#> 
#>    Stratum observations: 0
#> 
#>    Community observations: 0
#> 
#>    Site observations: 25
#> 
#>    Surface cover observations: 0
#> 
#> ================================================================

head(showElementTable(x, "siteObservation"))

#>    plotName obsStartDate soil_1_method soil_1_value
#> 1    LGM08r   2011-02-17            pH            7
#> 2 LGM08r_1Q   2011-02-17            pH            7
#> 3 LGM08r_2Q   2011-02-17            pH            7
#> 4 LGM08r_3Q   2011-02-17            pH            7
#> 5 LGM08r_4Q   2011-02-17            pH            7
#> 6    LGM16l   2011-02-15            pH            7