Fills the information for a given research project

Fills the information for a given research project

fillProjectInformation(target, title, personnel = list(),
  abstract = "", funding = "", studyAreaDescription = "",
  designDescription = "", citationString = "", DOI = "",
  verbose = TRUE)

Arguments

target
The initial object of class VegX to be modified.

title
A string with the title of the project. It can match one of the existing project or a new one will be defined.

personnel
A named list where element names are roles and values are party names (e.g. list(contributor = "John Smith")).

abstract
A string summarizing the aims and findings of the project.

funding
A string with information about funding agencies.

studyAreaDescription
A string describing the physical area associated with the research project, potentially including coverage, climate, geology, distrubances, etc..

designDescription
A string describing the overall plot placement design.

citationString
A string of the bibliographic reference of a document describing the project.

DOI
A string with the DOI the resource related to citationString.

verbose
A boolean flag to indicate console output of the data integration process.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

x = fillProjectInformation(newVegX(), "MOKIHINUI HYDRO PROPOSAL - LOWER GORGE 2011",
           personnel = c(contributor = "Susan K. Wiser"),
           abstract = paste("Characterise the forest and riparian vegetation",
                            " in the lower Mokihinui gorge, and compare this",
                            " with the vegetation in (a) North Branch gorge of Mokihinui",
                            " and (b) Karamea catchment."),
           studyAreaDescription = "Mokihinui and Karamea catchments. Forest riparian habitat.")

#>  1 new party(ies) added to the document as individuals.