Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 0cd141b7-fb6a-4c93-9e8d-b90475feb5da            <NA>
#> 2  LGM08r_1Q 9a2db894-0e24-4eb8-9f0d-8319144d996e          LGM08r
#> 3  LGM08r_2Q 1b3eca35-30d7-442d-8c8a-085112ca8021          LGM08r
#> 4  LGM08r_3Q a0f1f03d-b5ef-4b13-8501-6f4d3282e598          LGM08r
#> 5  LGM08r_4Q ea469f12-c881-47c3-8179-ba14b9c282f1          LGM08r
#> 6     LGM16l 70a05a38-7f94-4a14-a2d4-2dc7cce52041            <NA>
#> 7  LGM16l_1Q b4c13ac5-0196-48de-8705-7513d0bcc52e          LGM16l
#> 8  LGM16l_2Q c4c78cf2-50e4-4eda-b1f9-62a9ce490a3a          LGM16l
#> 9  LGM16l_3Q 959a6da3-fefa-436d-804f-ce2d40fb8513          LGM16l
#> 10 LGM16l_4Q a0d56295-a781-465c-bc8b-2bc646e14bdf          LGM16l
#> 11    LGM38h 3b54eedf-f7f4-4367-8d61-932dc77bdfdd            <NA>
#> 12 LGM38h_1Q 932ab663-00b3-4487-87f1-8adba044fad0          LGM38h
#> 13 LGM38h_2Q 9a5fa5b6-7748-4931-80c9-e52a098231a7          LGM38h
#> 14 LGM38h_3Q 948f3a8b-292c-4cab-b64f-051eed1574da          LGM38h
#> 15 LGM38h_4Q 89b68ba4-ad02-4bf3-ab75-3c4e095a1f33          LGM38h
#> 16    LGM45h 085384b0-166a-471f-9b03-e1cbd1b25323            <NA>
#> 17 LGM45h_2Q 54129bb9-75e3-4bb9-9c46-d08e3852b4d2          LGM45h
#> 18 LGM45h_3Q cfadbd6f-841a-4019-b48f-f401eafe6b99          LGM45h
#> 19 LGM45h_4Q 3007a34c-4e5c-4afa-893c-2ab141a31137          LGM45h
#> 20    LGM43h 1d6ff242-f439-4aac-bb0d-ad2a39c34fcc            <NA>
#> 21 LGM43h_1Q 6d94971e-307b-4905-a386-3e8a88fc6aea          LGM43h
#> 22 LGM43h_2Q b4a8bda1-3f57-47c9-845d-87feb3e8c1fd          LGM43h
#> 23 LGM43h_3Q 82ce95ac-e841-4f61-ace8-40edefcf7069          LGM43h
#> 24 LGM43h_4Q 324874cd-6162-4208-9738-fab69ee51e83          LGM43h
#> 25 LGM45h_1Q 6f294bf2-c0af-46c6-9d42-4126ba90fe45          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot