Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 84f5e1e1-4b7e-4dfd-bb76-ae9f06f80dc8            <NA>
#> 2  LGM08r_1Q edc019a1-952d-432d-9e35-c8040f1c84a9          LGM08r
#> 3  LGM08r_2Q 9fc4198a-1f3d-4874-a3c4-8392525fe8b5          LGM08r
#> 4  LGM08r_3Q 9524347a-285e-4ed9-8e47-f66edb1cc25c          LGM08r
#> 5  LGM08r_4Q 7a164e8e-69f9-4e31-91d2-4ef4184ecb38          LGM08r
#> 6     LGM16l b962b49e-86e0-4d5b-a4bb-87fdd9dab644            <NA>
#> 7  LGM16l_1Q 7052bc07-b1bc-4e14-aea9-9ae15b0a4e94          LGM16l
#> 8  LGM16l_2Q 02db8c65-b75d-4b95-ae6f-7fe5e60769e9          LGM16l
#> 9  LGM16l_3Q 64d75930-097e-42de-ad13-426802210570          LGM16l
#> 10 LGM16l_4Q 9ac209c8-269f-41c4-9d7a-13bc735b260f          LGM16l
#> 11    LGM38h 6a650d02-09b0-44e2-9ec8-8b7a655b1118            <NA>
#> 12 LGM38h_1Q 1178c626-830c-4604-841f-bfa7d864bb5e          LGM38h
#> 13 LGM38h_2Q 70a3713f-2954-4c6d-afcb-586b38e4cc6f          LGM38h
#> 14 LGM38h_3Q fe883cae-b4b4-452b-acb1-9627b5dd0fc6          LGM38h
#> 15 LGM38h_4Q 19c8159a-4e4b-4dfc-a832-7daa9908d69d          LGM38h
#> 16    LGM45h dbab908f-2e31-43c5-b55d-f155a5ad707f            <NA>
#> 17 LGM45h_2Q bb052d43-8d19-447c-85d4-9045af7c02c5          LGM45h
#> 18 LGM45h_3Q c3d1e6cc-4c89-4080-b79f-1ecde6f3be09          LGM45h
#> 19 LGM45h_4Q 59e2de75-5325-445d-8522-1fa4f1b6e3bf          LGM45h
#> 20    LGM43h d7009d91-8ee1-4b6a-ab7f-f7cca8ce5cbd            <NA>
#> 21 LGM43h_1Q cd4d24a7-f00f-45dc-a5f9-4d38190e0ea6          LGM43h
#> 22 LGM43h_2Q aea1dfcc-4056-436c-9b8a-8f1a278b128c          LGM43h
#> 23 LGM43h_3Q 7ef61a9b-b0bf-45c5-a587-7ea6cec7055d          LGM43h
#> 24 LGM43h_4Q 8aa557c7-9214-4648-b2be-15456fadc484          LGM43h
#> 25 LGM45h_1Q e09af880-39da-4886-94ed-619b95050b82          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot