Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r a3f9f288-2ae1-44fe-8476-f76cf97933da            <NA>
#> 2  LGM08r_1Q 04f903f4-0f86-47e2-b29c-df1d005107c6          LGM08r
#> 3  LGM08r_2Q 93053a6b-a87b-4e6c-9a8d-dbb25347990b          LGM08r
#> 4  LGM08r_3Q dd2849d2-a297-4a10-95e9-0a66709d7e5e          LGM08r
#> 5  LGM08r_4Q 4190722e-da9a-46e8-8b32-9f2ab4f7d13f          LGM08r
#> 6     LGM16l 80ac08a1-31c1-483a-bdfa-4c8d8dcbf5ee            <NA>
#> 7  LGM16l_1Q c341f9e6-4506-482b-b546-eadeacbf4550          LGM16l
#> 8  LGM16l_2Q 0adc6c20-551f-4806-a083-ab9b909aa232          LGM16l
#> 9  LGM16l_3Q 5031bd30-c3cb-4ba3-a00d-159f8b4f049a          LGM16l
#> 10 LGM16l_4Q aa4e25f0-355c-46b2-895f-e6845c07f873          LGM16l
#> 11    LGM38h 6eccfc19-ce22-4661-a879-94c7e5d3df38            <NA>
#> 12 LGM38h_1Q 197a8e78-5d57-4043-a60f-76727e51e07d          LGM38h
#> 13 LGM38h_2Q 7d54138b-76f6-4133-bacf-e787a381005d          LGM38h
#> 14 LGM38h_3Q b84471d7-6523-49d1-813a-9b03ba5757c5          LGM38h
#> 15 LGM38h_4Q d5f68862-c0e9-4a45-9909-82fb2e8ae759          LGM38h
#> 16    LGM45h 27f0377a-bda7-49a5-82d4-42cc0966e50c            <NA>
#> 17 LGM45h_2Q e3f43219-9c47-4c96-b244-88e406216de2          LGM45h
#> 18 LGM45h_3Q 2340e74f-fb04-46f4-a547-68cf68af0303          LGM45h
#> 19 LGM45h_4Q 511ddb98-d253-45e5-a195-53d64ce08723          LGM45h
#> 20    LGM43h 005b4f78-fce4-4713-b751-7dd73eb3b30b            <NA>
#> 21 LGM43h_1Q 3820ef7d-b94c-4c5e-9a9c-f54e2b4dd480          LGM43h
#> 22 LGM43h_2Q a816c768-615c-42df-949a-124eac6002c3          LGM43h
#> 23 LGM43h_3Q 265ef64f-bf12-4fac-95d8-619eabca554b          LGM43h
#> 24 LGM43h_4Q 35ca597d-3a12-4c28-851a-ff28629847dc          LGM43h
#> 25 LGM45h_1Q dbdf0682-b6ae-4162-871a-5497d2e78d92          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot