Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot",
  replace.existing = FALSE)

Arguments

target
The initial object of class VegX to be modified

element
The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing
A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)

#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")

#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")

#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 07b81059-8c1b-41a8-8c09-acc497e346da            <NA>
#> 2  LGM08r_1Q f4e774b7-8275-4890-897a-16a3c54777f1          LGM08r
#> 3  LGM08r_2Q 12f4b47c-f313-4096-8632-bbafc5dd0458          LGM08r
#> 4  LGM08r_3Q 9907180f-127a-4ea9-bf44-4214d81037ef          LGM08r
#> 5  LGM08r_4Q a6ea5dd1-66bb-4cf2-8d8d-240a9c9a957d          LGM08r
#> 6     LGM16l 13654713-f980-4438-bccd-fc770e826812            <NA>
#> 7  LGM16l_1Q f92a0489-3315-479f-9803-65791dd3eb6b          LGM16l
#> 8  LGM16l_2Q 74ce34aa-a8e8-476b-83e2-3b50bcafa149          LGM16l
#> 9  LGM16l_3Q 2460cea4-f441-490a-9de0-0dc48526d803          LGM16l
#> 10 LGM16l_4Q 2d19d622-1cd1-4424-a44a-72721daf50d0          LGM16l
#> 11    LGM38h 4da51737-ef45-46b9-b74d-e95186d765f8            <NA>
#> 12 LGM38h_1Q 48a0f3b2-d103-4327-98d1-f00bf1f9bf09          LGM38h
#> 13 LGM38h_2Q fb11d6ae-6bbc-4e5c-b18a-3ad4c07d89c8          LGM38h
#> 14 LGM38h_3Q a3d3ed7f-1fd0-4c38-8a90-b4438387afed          LGM38h
#> 15 LGM38h_4Q 5fbad08f-854c-4609-acb3-d5363485c38f          LGM38h
#> 16    LGM45h fbe99399-8d7f-42fc-9b41-380b4a26ee77            <NA>
#> 17 LGM45h_2Q 504e2621-753b-422a-b923-8b7fbcb67e19          LGM45h
#> 18 LGM45h_3Q 32dc2aad-ede4-4ac4-b119-325017bf57a0          LGM45h
#> 19 LGM45h_4Q 010556a4-e992-4a37-b040-8e59510f9a18          LGM45h
#> 20    LGM43h e1337a02-f41a-47bc-8415-3cbc8376a24f            <NA>
#> 21 LGM43h_1Q f5a50529-d62f-4e72-a999-a1d21a2e34d2          LGM43h
#> 22 LGM43h_2Q d202eddb-d7a2-4f1c-b955-b3690ed746cc          LGM43h
#> 23 LGM43h_3Q 0a2dc9d2-d3f0-4352-b133-78bcbe287b1f          LGM43h
#> 24 LGM43h_4Q 312f1419-0431-4cd8-b8a4-06595a70bf32          LGM43h
#> 25 LGM45h_1Q 1e8d5684-2ccd-42f9-af51-8eb21c597911          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot