Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 5b8b1e09-43cc-4758-ab74-8ddd21952059            <NA>
#> 2  LGM08r_1Q e84612bb-48ef-4e70-b51e-0be0ef5d88d5          LGM08r
#> 3  LGM08r_2Q 7fa4a35a-54fb-4515-af26-47427ee162dd          LGM08r
#> 4  LGM08r_3Q 381bc73a-9ae7-416a-abdc-8331789a8c48          LGM08r
#> 5  LGM08r_4Q 59c1e556-358e-439a-9759-f5f3a4c49121          LGM08r
#> 6     LGM16l 5d996ff0-3b71-4a09-84f9-64b6da098b20            <NA>
#> 7  LGM16l_1Q 1563f46d-7d20-44ff-9c3a-ee1178d3d6c5          LGM16l
#> 8  LGM16l_2Q efa42e19-ed3c-493b-b18a-08531d07bf91          LGM16l
#> 9  LGM16l_3Q 54fcb42f-4048-4e3e-8703-3ae811ae2755          LGM16l
#> 10 LGM16l_4Q d97db2ad-60bf-42ff-87c8-019b18b85d7b          LGM16l
#> 11    LGM38h e08bd16b-ac71-411d-a501-656b75d77f27            <NA>
#> 12 LGM38h_1Q e6db8043-9509-47e8-8e7c-9daa397b3523          LGM38h
#> 13 LGM38h_2Q d443ade0-b39d-421e-bade-61ecebd02017          LGM38h
#> 14 LGM38h_3Q a016cd5c-4817-430d-b522-0cc249bcfae1          LGM38h
#> 15 LGM38h_4Q 799682a9-cb83-40b2-a5cc-d56d44dd9037          LGM38h
#> 16    LGM45h dcd06141-5b12-40f9-b514-28037e14e5fa            <NA>
#> 17 LGM45h_2Q 57c6ed95-9472-4367-967e-d1367800ecad          LGM45h
#> 18 LGM45h_3Q acc50af3-ad98-4726-ab89-43257cdb5807          LGM45h
#> 19 LGM45h_4Q e8c5d975-37d8-41c7-ac8b-4dc55e87ff8b          LGM45h
#> 20    LGM43h 6402f393-a65a-4869-8754-565a43eab893            <NA>
#> 21 LGM43h_1Q 26292f07-329f-4102-9b65-8510dca19f34          LGM43h
#> 22 LGM43h_2Q bfa4ff5f-9b3e-4381-a06e-bba21735895e          LGM43h
#> 23 LGM43h_3Q 4b70f1b4-ac7b-4322-9c75-60eca2725bc7          LGM43h
#> 24 LGM43h_4Q 5afec668-6617-4140-9995-3e94c663d77f          LGM43h
#> 25 LGM45h_1Q f480f7fb-b5a3-42ce-af57-ac9298ca7703          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot