Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r b09b1f02-ca4a-494e-b928-196673c2196a            <NA>
#> 2  LGM08r_1Q 2b24ab8c-1f17-421e-b1e3-be64768fa8dc          LGM08r
#> 3  LGM08r_2Q 82202100-ecdf-4d4b-b868-5d98f2cb87f2          LGM08r
#> 4  LGM08r_3Q fbaf63e0-cd70-4291-a430-3de1a0229863          LGM08r
#> 5  LGM08r_4Q 1d5e6ea5-e48c-4d5d-83b8-6b9960916444          LGM08r
#> 6     LGM16l 4a6ac9ad-7d33-4467-b9dc-285e98fca56b            <NA>
#> 7  LGM16l_1Q 6b795666-2680-437f-9b0d-cb5388b70480          LGM16l
#> 8  LGM16l_2Q b5424f03-0ae8-42c6-8c73-631cc8247d23          LGM16l
#> 9  LGM16l_3Q 2394aa13-ad9a-4e8a-80e3-839d8a596148          LGM16l
#> 10 LGM16l_4Q 73d5c28e-ad87-40dd-80dc-9b326987d4ce          LGM16l
#> 11    LGM38h 258e6aa5-d526-4e8c-ae10-5929b012a2cf            <NA>
#> 12 LGM38h_1Q 659c3048-4949-40cd-b902-3e596e12989f          LGM38h
#> 13 LGM38h_2Q 30eb34d5-9757-4113-8ae8-3c825a350535          LGM38h
#> 14 LGM38h_3Q 4d7c6165-464b-49af-85cd-4fdd776b47af          LGM38h
#> 15 LGM38h_4Q 4d1ccb3f-ee7a-4434-bb2b-8d184099ef4b          LGM38h
#> 16    LGM45h b82b6ddb-a2a1-4452-a6dc-ef672a5729fd            <NA>
#> 17 LGM45h_2Q 7ede0229-6e06-457a-833a-1397071f0c7a          LGM45h
#> 18 LGM45h_3Q ecb541eb-b6bb-42e0-b525-29e4e33c43bf          LGM45h
#> 19 LGM45h_4Q e4a132fd-6fc2-4a55-80bd-af7418dec81b          LGM45h
#> 20    LGM43h 1f258573-7b7b-400d-9c24-a536d31d4724            <NA>
#> 21 LGM43h_1Q 5b0b896d-b18b-4918-a4cd-b7ab90640a93          LGM43h
#> 22 LGM43h_2Q bcca36d2-e204-443b-b649-ad29f2a26bcf          LGM43h
#> 23 LGM43h_3Q 24fb4e3f-253b-4793-b6e4-a2014fed7c32          LGM43h
#> 24 LGM43h_4Q 6501da09-eeb9-4e0c-85e6-68998656bdb6          LGM43h
#> 25 LGM45h_1Q 956c7b06-ddb6-46d0-bde6-4aa1227dd8e0          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot