Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 3fc6b2fb-cba0-4d02-8b2e-2bbff58e34fd            <NA>
#> 2  LGM08r_1Q b0a41b4d-4e27-4eb1-bcf8-d5e27f5eca67          LGM08r
#> 3  LGM08r_2Q 2100b8df-5e5a-4af3-8e20-adb5d2c9e921          LGM08r
#> 4  LGM08r_3Q 35daa3dd-e42e-47c8-9fe1-c47cfea1c59a          LGM08r
#> 5  LGM08r_4Q 93a7eb5b-3332-47c0-9b13-ef391684f8f0          LGM08r
#> 6     LGM16l e5cef232-d50d-4a83-97c5-f843220a2954            <NA>
#> 7  LGM16l_1Q c967be75-6847-4161-a077-2990fa931b40          LGM16l
#> 8  LGM16l_2Q 91da2ecb-e577-4213-9f01-a79094b9fb50          LGM16l
#> 9  LGM16l_3Q 710e97f5-9a6e-4f14-9fbc-11e4559f5f37          LGM16l
#> 10 LGM16l_4Q 7dc9ea68-8bbc-482b-9699-efaac6f1aeaa          LGM16l
#> 11    LGM38h 8ea9e758-66d3-42b3-a716-b374ee275770            <NA>
#> 12 LGM38h_1Q 4d9380ff-b790-48e6-b046-e7f19427295a          LGM38h
#> 13 LGM38h_2Q 3ba34e63-9eb6-47a5-b233-93b02b74f460          LGM38h
#> 14 LGM38h_3Q ed79a836-40ac-4116-accf-b0ff56a60bc7          LGM38h
#> 15 LGM38h_4Q 4c8defba-f673-43dc-b2ea-27c0b46dad58          LGM38h
#> 16    LGM45h 1dd1a39a-c82d-4305-923b-930649db1efe            <NA>
#> 17 LGM45h_2Q 7614e3d6-665b-4bb9-bb53-09550b204fab          LGM45h
#> 18 LGM45h_3Q 4355453d-2a9c-413d-a6f6-49fab723a059          LGM45h
#> 19 LGM45h_4Q 8f889ecb-e80d-4b11-bd43-3aa0464db95f          LGM45h
#> 20    LGM43h f120b18a-53f7-49ab-a0a5-5c142e2017d1            <NA>
#> 21 LGM43h_1Q 0476f464-8ce8-4893-b53a-5aa7e053dda5          LGM43h
#> 22 LGM43h_2Q f02e4959-aa61-45d5-92ef-0c15bc5f5fac          LGM43h
#> 23 LGM43h_3Q f6bc8f2a-58ad-4772-83a4-d1a01ce4112a          LGM43h
#> 24 LGM43h_4Q 0d27714d-eba1-47a9-9745-f2aa8fa8c684          LGM43h
#> 25 LGM45h_1Q f41a92eb-8ae0-4be3-a587-94cd4aae21f9          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot