Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 2b44aa71-4271-4cdc-b1cc-5be4d9c984e2            <NA>
#> 2  LGM08r_1Q b51443fc-d843-41f9-894a-641ca5f812a6          LGM08r
#> 3  LGM08r_2Q f82114b5-bd48-4ff3-9503-5a93eb14ff6d          LGM08r
#> 4  LGM08r_3Q d751b2a5-496a-42c7-b8f8-dbe1d0f88152          LGM08r
#> 5  LGM08r_4Q 8c00a262-e713-433d-ab9f-31ada0fbdbbc          LGM08r
#> 6     LGM16l a0db946c-8c45-458d-aa98-3058164c2759            <NA>
#> 7  LGM16l_1Q 0da1eff9-e5b5-47f4-98b8-d497ae4a32ad          LGM16l
#> 8  LGM16l_2Q 17a45233-bea5-4094-ad04-cdaecf2c89a4          LGM16l
#> 9  LGM16l_3Q 47c30fc4-869b-4cbb-9eea-428cc3928deb          LGM16l
#> 10 LGM16l_4Q 36ad6c16-cbe3-4eae-a6e2-b86bb94859c0          LGM16l
#> 11    LGM38h 183e14fb-5945-4ff1-b303-d15842e30a3f            <NA>
#> 12 LGM38h_1Q 3cce8f12-a035-4005-8b90-7b2057d17f94          LGM38h
#> 13 LGM38h_2Q dd1ef4fd-0617-4849-9830-32557ce66310          LGM38h
#> 14 LGM38h_3Q e5fdd8b2-07d5-4e34-9a12-0f346f6f3c99          LGM38h
#> 15 LGM38h_4Q 3ca66a0c-9966-42a1-8ebe-b1a14af31d05          LGM38h
#> 16    LGM45h 295c3e78-cd28-49d4-81f8-6e3b63d83c5e            <NA>
#> 17 LGM45h_2Q fe77fb22-3873-4490-8816-1e60328e0148          LGM45h
#> 18 LGM45h_3Q 72f6d0bd-7f43-449a-881c-db0adc11c527          LGM45h
#> 19 LGM45h_4Q 0cba5533-601b-4d8f-b525-611c44bdaad3          LGM45h
#> 20    LGM43h b5715c10-6e84-494c-9f2e-c52b4d627d23            <NA>
#> 21 LGM43h_1Q d045dce1-2ba6-424b-b615-5ac2817beb72          LGM43h
#> 22 LGM43h_2Q 33720ce7-40c1-4046-98e7-9b37395cf56c          LGM43h
#> 23 LGM43h_3Q a98d7d86-342c-4204-a959-56263241c4b4          LGM43h
#> 24 LGM43h_4Q ce1e6b2a-2653-4b5a-a6d9-5817e879bf8a          LGM43h
#> 25 LGM45h_1Q 183fee44-5d24-4edd-b345-b8d3bb968abc          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot