Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r cf09bffd-9560-4fc0-a322-d1dd50bcb53a            <NA>
#> 2  LGM08r_1Q 9e4b9622-41e1-4265-b6b7-30f7943d6942          LGM08r
#> 3  LGM08r_2Q 72d8b57c-d7bf-4232-8c64-6e97a4ba4cdc          LGM08r
#> 4  LGM08r_3Q 55d8b602-ce06-4f2e-87c5-4f963119a604          LGM08r
#> 5  LGM08r_4Q ec0d3beb-9f10-49b1-ad66-1f6714591bee          LGM08r
#> 6     LGM16l 7bd7306d-2e3e-4202-bf36-154290e08802            <NA>
#> 7  LGM16l_1Q ddad3c73-d2f5-4e90-9476-3c8e3b578cf5          LGM16l
#> 8  LGM16l_2Q 51d6c597-ffc3-4347-91b5-fbc6acd6e104          LGM16l
#> 9  LGM16l_3Q a3b5778e-aa33-4389-a6c1-53633c8c6083          LGM16l
#> 10 LGM16l_4Q 724ea9c2-e321-42bb-a5a4-46ae250344d9          LGM16l
#> 11    LGM38h 119a27b2-6580-4dab-adf9-a7b83ca72de2            <NA>
#> 12 LGM38h_1Q e5e5fdb9-a441-4a02-bb95-d6a00a7a893c          LGM38h
#> 13 LGM38h_2Q 02f058f7-7d02-4fe9-9429-bff5fcaa4160          LGM38h
#> 14 LGM38h_3Q c4a6214e-edad-41b9-b52b-ee81e323519b          LGM38h
#> 15 LGM38h_4Q b105c3e7-f945-4fa2-be14-1597ab91b890          LGM38h
#> 16    LGM45h b59d2656-beb0-4f5f-8e7a-4792a2842af4            <NA>
#> 17 LGM45h_2Q d4d6e2aa-ed8e-4b44-b6aa-a4d6534bcbf0          LGM45h
#> 18 LGM45h_3Q 0f02c62b-94c3-4b28-8261-5b083dfc614a          LGM45h
#> 19 LGM45h_4Q a14883e6-cfb4-4624-b986-3548465dfdd9          LGM45h
#> 20    LGM43h 5b4b0829-b6e0-4727-94ce-34e271d41800            <NA>
#> 21 LGM43h_1Q f2c6367a-74c4-4c85-8795-1b111f3aa27a          LGM43h
#> 22 LGM43h_2Q b067d48c-64c9-4450-aeba-2415a26a73b2          LGM43h
#> 23 LGM43h_3Q e1d1af49-c803-480a-b37c-022a343db3b6          LGM43h
#> 24 LGM43h_4Q 3797516c-5022-4eac-a1d9-c9a8f50cd657          LGM43h
#> 25 LGM45h_1Q 00cf9885-d7f8-4696-a832-26064aaaf378          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot