Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 99cf08a9-b68a-4443-bfdd-90ae72398650            <NA>
#> 2  LGM08r_1Q b32cbe58-bdfd-4e23-a7a1-a7d23e2b3008          LGM08r
#> 3  LGM08r_2Q fbad2add-df48-410e-a0ce-717b5112761c          LGM08r
#> 4  LGM08r_3Q df81eefd-ee12-4d59-b397-24dc823a4322          LGM08r
#> 5  LGM08r_4Q 749e559c-83d1-4c93-afda-175c8df37f7a          LGM08r
#> 6     LGM16l 31f99034-6282-4a01-aba7-03afdded1f38            <NA>
#> 7  LGM16l_1Q fcafc8bc-c9c5-4262-b341-7ed40751f470          LGM16l
#> 8  LGM16l_2Q 36fc09c8-cb01-4cbe-aee5-6a0383981486          LGM16l
#> 9  LGM16l_3Q afe4f273-aabc-4547-a573-ffb6dcf3dbf3          LGM16l
#> 10 LGM16l_4Q ca8e8896-5e8f-4393-a9e3-20befcd8a9dc          LGM16l
#> 11    LGM38h e0c4aeff-1358-40cf-a639-404b2aef28a8            <NA>
#> 12 LGM38h_1Q a0d4e7de-191a-46c3-a27f-927951f921e5          LGM38h
#> 13 LGM38h_2Q 40362b70-84e6-4ae2-9405-d49a27e30aec          LGM38h
#> 14 LGM38h_3Q 157d349d-0f51-4046-97e0-fa60f9485079          LGM38h
#> 15 LGM38h_4Q 20ec2c2c-e507-4e5f-a2a6-2532c319d8b7          LGM38h
#> 16    LGM45h 1b2d0f2b-9a2d-4ae6-a9db-4f6c35b68c54            <NA>
#> 17 LGM45h_2Q b818c7bb-ebcf-4292-8771-d1021706c901          LGM45h
#> 18 LGM45h_3Q 0e1eeae4-4c04-496a-a56f-8016b5c0dc60          LGM45h
#> 19 LGM45h_4Q 4a4469f9-2db9-4a6f-9bf4-0dc5eb8d2172          LGM45h
#> 20    LGM43h 19b35433-7b43-4bc2-9470-85061cc439af            <NA>
#> 21 LGM43h_1Q 2b923f51-5ded-4682-be95-8488fe369db3          LGM43h
#> 22 LGM43h_2Q d953b4cd-d14e-41d9-a47b-8cf9a09320e7          LGM43h
#> 23 LGM43h_3Q cb1ab997-95e2-4034-990e-0696def13d39          LGM43h
#> 24 LGM43h_4Q 6ecbe7e8-3611-4669-b998-01201d57c57a          LGM43h
#> 25 LGM45h_1Q 9119c095-1180-4b9b-affa-b29e15a91883          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot