Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r e785d558-c909-41f5-8666-c958f824f3a7            <NA>
#> 2  LGM08r_1Q e9247578-3758-4e12-a736-995b56ad0e47          LGM08r
#> 3  LGM08r_2Q cd672d4c-28dd-4e83-9d92-9ab80ded76ab          LGM08r
#> 4  LGM08r_3Q aaeaa597-0016-41bd-b470-c7ae43deb93b          LGM08r
#> 5  LGM08r_4Q e413e763-5587-4444-9ad9-e5c43d77ee27          LGM08r
#> 6     LGM16l 9aa6bab0-3589-45da-8a99-c8f1ee2896db            <NA>
#> 7  LGM16l_1Q df8e87c7-b2db-4805-9589-70f5b080b04e          LGM16l
#> 8  LGM16l_2Q d8cf614e-e75f-4fb3-8d46-2ce3986579e5          LGM16l
#> 9  LGM16l_3Q febd5d38-56d6-43d1-8608-efe13502f789          LGM16l
#> 10 LGM16l_4Q bf2b12c1-69b8-40e9-91d7-1c2b8c26dda6          LGM16l
#> 11    LGM38h f68291fe-c4bd-4502-840a-d02dbc036d84            <NA>
#> 12 LGM38h_1Q ccf08d98-d65e-4037-876b-032e2027eb3e          LGM38h
#> 13 LGM38h_2Q 7c4d29f3-6db8-4ddf-b8c0-02805f2f5dad          LGM38h
#> 14 LGM38h_3Q e008c46e-c332-4236-a111-2797444b8e03          LGM38h
#> 15 LGM38h_4Q 1358c286-da94-4081-8e50-ec4943d3b025          LGM38h
#> 16    LGM45h b7b5b68a-cf2e-4341-b8aa-fd76950d7049            <NA>
#> 17 LGM45h_2Q c0a07b43-7b59-4ab6-9868-08917e46f6ef          LGM45h
#> 18 LGM45h_3Q 56c488c6-2467-4fcf-ae81-9a0c3317d97d          LGM45h
#> 19 LGM45h_4Q 8906cbb6-125d-44f2-884d-f30d90942e67          LGM45h
#> 20    LGM43h b4cd547b-c75e-470b-9bb2-4a5b6324f05b            <NA>
#> 21 LGM43h_1Q aa4ccf42-697e-4aec-ac71-d10459fc5ba9          LGM43h
#> 22 LGM43h_2Q b3b5a548-308c-41f9-9a95-f5a52d6bbac8          LGM43h
#> 23 LGM43h_3Q ccf1fb13-d366-4b50-a620-da354567a84e          LGM43h
#> 24 LGM43h_4Q 26235ab7-c7a7-4d52-9202-39b21014062c          LGM43h
#> 25 LGM45h_1Q 0e79111f-b738-44c7-8303-fd8cfd550b39          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot