Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot",
  replace.existing = FALSE)

Arguments

target
The initial object of class VegX to be modified

element
The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing
A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)

#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")

#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")

#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 4477332f-f40b-4fe1-9e4f-26a2d24e7110            <NA>
#> 2  LGM08r_1Q 6d704e04-81d8-4073-9c1f-79c6d59b6cf1          LGM08r
#> 3  LGM08r_2Q cca677fa-ce6b-4f17-a937-597fa0b0f15e          LGM08r
#> 4  LGM08r_3Q 57097c2c-0957-4df5-ad24-0a943c9ef899          LGM08r
#> 5  LGM08r_4Q 80449e0c-7855-4377-b6c1-8518c7d77369          LGM08r
#> 6     LGM16l a4c01e63-5b40-4d0f-a37e-3a885f6a7d07            <NA>
#> 7  LGM16l_1Q 08baa851-3f98-4748-b8d9-0bf418eb6ae4          LGM16l
#> 8  LGM16l_2Q 2b9d588b-88a9-44e1-b0b4-df6a86d222a6          LGM16l
#> 9  LGM16l_3Q 56911136-1a47-4ca3-99f3-8098e55d29bc          LGM16l
#> 10 LGM16l_4Q 7042c42d-736a-45bc-9518-5c82dc2b734f          LGM16l
#> 11    LGM38h b88c3d77-c8e2-4fa5-8c8e-15ecb600a5ad            <NA>
#> 12 LGM38h_1Q f93a1360-4500-4942-b23d-21344452b14f          LGM38h
#> 13 LGM38h_2Q f1a37437-bf56-46fd-b77e-6bc8d77f41db          LGM38h
#> 14 LGM38h_3Q 6aea0ebc-f1ba-4cbe-9946-1ae53fadfc32          LGM38h
#> 15 LGM38h_4Q c65de31d-d49a-4363-94a4-d9d89b6b80fc          LGM38h
#> 16    LGM45h 905aed61-986c-48d0-9833-610f26f130f2            <NA>
#> 17 LGM45h_2Q 8f2f7f38-732c-4207-8cbc-02c5f11a4c5b          LGM45h
#> 18 LGM45h_3Q 9b4204d9-d490-430f-a9d9-07cede59ea85          LGM45h
#> 19 LGM45h_4Q 580f8f13-9229-400f-9a5d-80876a3527ec          LGM45h
#> 20    LGM43h ba60ee0d-8f37-4f07-a166-ab2efbaee6d4            <NA>
#> 21 LGM43h_1Q bdd41061-b0dd-4093-b0e8-80e0463c781c          LGM43h
#> 22 LGM43h_2Q 819f250b-cf95-4a55-be80-fa2020485775          LGM43h
#> 23 LGM43h_3Q 600acb52-50df-43ae-b8ef-0198ce87d12d          LGM43h
#> 24 LGM43h_4Q 8670e5f3-9722-4935-ae61-40b53275cbd1          LGM43h
#> 25 LGM45h_1Q 48c7b485-9027-41b1-a7e7-817a520bd156          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot