Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 98d8d6ab-8e5e-40bb-bca9-8f46ba979b2b            <NA>
#> 2  LGM08r_1Q 872f683a-5800-4e7a-b194-948b993c27c5          LGM08r
#> 3  LGM08r_2Q 7771f866-6293-4afa-8b27-3c6d3d988e17          LGM08r
#> 4  LGM08r_3Q 6256d824-9a8f-43fc-9db8-b4313f0e67a8          LGM08r
#> 5  LGM08r_4Q 0aaeee53-f704-4252-8d65-bcfe415e36a9          LGM08r
#> 6     LGM16l 839b8dea-185a-44ca-bda1-306871af5e6f            <NA>
#> 7  LGM16l_1Q ee1e6906-fe58-4c1a-b33d-008fb5dc34a0          LGM16l
#> 8  LGM16l_2Q 2c20d8b9-6b68-4303-96f3-6136381f3253          LGM16l
#> 9  LGM16l_3Q 0863debb-f37e-46f4-a886-a33e429c54bf          LGM16l
#> 10 LGM16l_4Q b3e0e84b-7f37-4d5f-a17e-98670f15cde5          LGM16l
#> 11    LGM38h ced6d667-88ff-49b2-8865-7cc7955dbfac            <NA>
#> 12 LGM38h_1Q 93fa5a7b-3060-4042-ad4b-50ec5dc8b00d          LGM38h
#> 13 LGM38h_2Q 910c8357-b587-4197-a2e0-19859acf4273          LGM38h
#> 14 LGM38h_3Q 556c6698-80b3-4c51-aae7-492cd604d5bb          LGM38h
#> 15 LGM38h_4Q 003282ca-f49a-42c2-9e77-7e28677e8f73          LGM38h
#> 16    LGM45h d4d28dd5-64eb-4f7b-b336-d777513a0abb            <NA>
#> 17 LGM45h_2Q 582b96a9-ae9c-43c0-82ff-7062059fb0ae          LGM45h
#> 18 LGM45h_3Q 9d105a98-3f70-4a3d-990c-14dbe91f9c04          LGM45h
#> 19 LGM45h_4Q a2b07fea-f8d7-4b3a-9a13-1ed2efe74d58          LGM45h
#> 20    LGM43h e0ee0491-5a6f-43a8-bb54-7a85ee4241a4            <NA>
#> 21 LGM43h_1Q d880cce8-6064-412b-a6d5-05b1d57360e4          LGM43h
#> 22 LGM43h_2Q 67191d8f-9067-48ac-af90-d7f209ca2393          LGM43h
#> 23 LGM43h_3Q 531bf757-4643-431c-869b-5e033355a404          LGM43h
#> 24 LGM43h_4Q 6907fcd6-350f-4d68-8590-156b4d9be5f4          LGM43h
#> 25 LGM45h_1Q 3f99f11e-4504-4314-98e8-0feda87cafc3          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot