Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

Usage

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r db0b4bb2-3c50-4b45-a20a-1b7634179f40            <NA>
#> 2  LGM08r_1Q 066dd359-9df1-400f-b492-ee127298b964          LGM08r
#> 3  LGM08r_2Q 9fb6c47c-49cf-465e-8f95-adf8e8a4f977          LGM08r
#> 4  LGM08r_3Q 1b6f03e1-1f76-46fc-a9dc-ae1512a2317b          LGM08r
#> 5  LGM08r_4Q 64b6b125-26c4-4232-802f-05d2ef7edcf8          LGM08r
#> 6     LGM16l efc024f2-31da-4b81-9e29-e0beed872e6c            <NA>
#> 7  LGM16l_1Q d8df77b8-a917-4b64-ab44-e49ce25e78d5          LGM16l
#> 8  LGM16l_2Q a2c6211e-0554-4e27-bfab-20d8544dc17b          LGM16l
#> 9  LGM16l_3Q e307c530-5141-4eb7-8de0-42d17b02c347          LGM16l
#> 10 LGM16l_4Q 79bad358-3fef-44a9-91ad-15c108f63194          LGM16l
#> 11    LGM38h f63117dd-f184-4eab-91e3-6ae85eb5bc66            <NA>
#> 12 LGM38h_1Q e7e7384b-d432-4492-8d4a-7d33d4859774          LGM38h
#> 13 LGM38h_2Q 76205500-bde3-4655-ae56-56dccc0fb3f6          LGM38h
#> 14 LGM38h_3Q 88f000aa-9edc-4fd3-bfa9-e74ef2072651          LGM38h
#> 15 LGM38h_4Q 563d5041-481e-46b7-9ac8-4a95e3b09b3d          LGM38h
#> 16    LGM45h 1a3ed3dd-97cc-4238-8bcb-87569847b342            <NA>
#> 17 LGM45h_2Q b39735b0-c395-4759-a2f9-3324b56c8356          LGM45h
#> 18 LGM45h_3Q 417775f3-f387-4a6b-baff-1bdcdd90ffd1          LGM45h
#> 19 LGM45h_4Q d86aa3dd-c8e4-4163-a69d-e2fb12598e9a          LGM45h
#> 20    LGM43h da71a8d5-2d03-4c99-b721-9dc0e985a32f            <NA>
#> 21 LGM43h_1Q af6a670c-ee3e-4af5-aad0-a8480ac72f27          LGM43h
#> 22 LGM43h_2Q 2918e1e6-99fd-4547-8b8e-8a0a731991e5          LGM43h
#> 23 LGM43h_3Q 13d20433-aa93-4985-8338-2b0bfd642943          LGM43h
#> 24 LGM43h_4Q 8fd7c4b1-a347-4fe4-9b3d-90e2f23a0d2d          LGM43h
#> 25 LGM45h_1Q d0eb7c53-f838-4966-a404-32c6753d2724          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot