Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot",
  replace.existing = FALSE)

Arguments

target
The initial object of class VegX to be modified

element
The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing
A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)

#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")

#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")

#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 063b7238-bca9-11e8-9f48-49df8e908e76            <NA>
#> 2  LGM08r_1Q 063b7239-bca9-11e8-9f48-49df8e908e76          LGM08r
#> 3  LGM08r_2Q 063b723a-bca9-11e8-9f48-49df8e908e76          LGM08r
#> 4  LGM08r_3Q 063b723b-bca9-11e8-9f48-49df8e908e76          LGM08r
#> 5  LGM08r_4Q 063b723c-bca9-11e8-9f48-49df8e908e76          LGM08r
#> 6     LGM16l 063b723d-bca9-11e8-9f48-49df8e908e76            <NA>
#> 7  LGM16l_1Q 063b723e-bca9-11e8-9f48-49df8e908e76          LGM16l
#> 8  LGM16l_2Q 063b723f-bca9-11e8-9f48-49df8e908e76          LGM16l
#> 9  LGM16l_3Q 063b7240-bca9-11e8-9f48-49df8e908e76          LGM16l
#> 10 LGM16l_4Q 063b7241-bca9-11e8-9f48-49df8e908e76          LGM16l
#> 11    LGM38h 063b7242-bca9-11e8-9f48-49df8e908e76            <NA>
#> 12 LGM38h_1Q 063dd474-bca9-11e8-9f48-49df8e908e76          LGM38h
#> 13 LGM38h_2Q 063dd475-bca9-11e8-9f48-49df8e908e76          LGM38h
#> 14 LGM38h_3Q 063dd476-bca9-11e8-9f48-49df8e908e76          LGM38h
#> 15 LGM38h_4Q 063dd477-bca9-11e8-9f48-49df8e908e76          LGM38h
#> 16    LGM45h 063dd478-bca9-11e8-9f48-49df8e908e76            <NA>
#> 17 LGM45h_2Q 063dd479-bca9-11e8-9f48-49df8e908e76          LGM45h
#> 18 LGM45h_3Q 063dd47a-bca9-11e8-9f48-49df8e908e76          LGM45h
#> 19 LGM45h_4Q 063dd47b-bca9-11e8-9f48-49df8e908e76          LGM45h
#> 20    LGM43h 063dd47c-bca9-11e8-9f48-49df8e908e76            <NA>
#> 21 LGM43h_1Q 063dd47d-bca9-11e8-9f48-49df8e908e76          LGM43h
#> 22 LGM43h_2Q 063dd47e-bca9-11e8-9f48-49df8e908e76          LGM43h
#> 23 LGM43h_3Q 063fb4b0-bca9-11e8-9f48-49df8e908e76          LGM43h
#> 24 LGM43h_4Q 063fb4b1-bca9-11e8-9f48-49df8e908e76          LGM43h
#> 25 LGM45h_1Q 063fb4b2-bca9-11e8-9f48-49df8e908e76          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot