Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r 70441491-3dd3-4e80-8d2b-e493c2cc0809            <NA>
#> 2  LGM08r_1Q 9b104ea3-5f1c-41ea-b5c8-6232ed4ea81f          LGM08r
#> 3  LGM08r_2Q 2a82a129-e37f-4b6a-9a2e-256cce60f871          LGM08r
#> 4  LGM08r_3Q b78ae8a3-7b97-4c80-9bf7-29692fcfa922          LGM08r
#> 5  LGM08r_4Q 35cb7f4a-8fb8-4269-be71-796a84afac93          LGM08r
#> 6     LGM16l 6cc70ff3-e03d-4cf7-96a2-1d281cab132a            <NA>
#> 7  LGM16l_1Q 535bf166-a851-4aa8-96fe-d8995d845b36          LGM16l
#> 8  LGM16l_2Q 3d218dc7-86a2-4aff-93f2-b1227eac560f          LGM16l
#> 9  LGM16l_3Q 85b6f4e8-3324-4f6c-84e0-a6df67aebf08          LGM16l
#> 10 LGM16l_4Q d7a920b5-b866-4c99-8b2a-a06752d5c640          LGM16l
#> 11    LGM38h 79cdf08d-c9ed-4679-b0e7-2e30e5e5f96f            <NA>
#> 12 LGM38h_1Q 7da92098-8e86-4944-994a-09adabee2ed0          LGM38h
#> 13 LGM38h_2Q f4efc4e4-1894-4429-bb7f-db7cd1fa555e          LGM38h
#> 14 LGM38h_3Q d9d0cb16-7a47-4dd5-8a1a-214c03e04a03          LGM38h
#> 15 LGM38h_4Q 607a47fc-e185-4d1d-83c7-453c5380ced1          LGM38h
#> 16    LGM45h 89690d49-d537-480e-92ac-b4a2015b25b6            <NA>
#> 17 LGM45h_2Q b17d8309-d0c0-473c-b2bc-7bcd97281179          LGM45h
#> 18 LGM45h_3Q 10ffb379-ae3b-4ad8-806d-0ff1457ba6e2          LGM45h
#> 19 LGM45h_4Q 396a09b1-4a63-4cc5-9a46-60d26fa02533          LGM45h
#> 20    LGM43h c6d3a760-5a01-458f-bb7e-c5c198763bf8            <NA>
#> 21 LGM43h_1Q eaac0443-5f1d-4411-b029-f83dc2246b20          LGM43h
#> 22 LGM43h_2Q a26cad6a-d452-4d20-a555-250f4c508774          LGM43h
#> 23 LGM43h_3Q 829e427b-f8ab-409d-9d98-24f320e4e8e1          LGM43h
#> 24 LGM43h_4Q 6d3857fb-ee08-4620-abbc-1f9cf45cf924          LGM43h
#> 25 LGM45h_1Q 99dc009b-3a86-48dd-b701-63511ada2b99          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot