Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r bb1815a1-ee81-4260-8134-09ed78396032            <NA>
#> 2  LGM08r_1Q da578ca5-8b7f-4dd4-91c3-b03af106401b          LGM08r
#> 3  LGM08r_2Q b3549eca-cfbd-41f1-b7f4-9ad97b433329          LGM08r
#> 4  LGM08r_3Q cb498a54-5d2b-4c9e-96f6-ed42f522f18e          LGM08r
#> 5  LGM08r_4Q b0fcbc90-d60a-46aa-a294-c5b5f004aa3c          LGM08r
#> 6     LGM16l 8328c558-ce68-41d8-9b03-19a85367b4e2            <NA>
#> 7  LGM16l_1Q 6047107f-c81e-4879-b08c-3e2df6f99f34          LGM16l
#> 8  LGM16l_2Q a362fe01-3d68-43f1-bcad-2a93104ee25c          LGM16l
#> 9  LGM16l_3Q 0ee8deef-414f-45d3-bef8-939f26320c40          LGM16l
#> 10 LGM16l_4Q 149b465b-7439-4712-b634-833fcee01d20          LGM16l
#> 11    LGM38h 72eabcfd-438f-4e3e-ab21-bcd10c03f405            <NA>
#> 12 LGM38h_1Q 81c98ac6-74a2-4f73-b9f6-031bba20ce43          LGM38h
#> 13 LGM38h_2Q 926b95db-a27c-4277-af73-ce5f88272d93          LGM38h
#> 14 LGM38h_3Q c5c4a924-b1e5-4f04-9930-f9944d45e699          LGM38h
#> 15 LGM38h_4Q a7064b9d-f892-4c2b-8daf-c0595ec31dd0          LGM38h
#> 16    LGM45h ff1fb175-acd9-49ba-8ca6-54129ba839d1            <NA>
#> 17 LGM45h_2Q fd4f6b61-9421-413d-85e7-51008dbe5f7a          LGM45h
#> 18 LGM45h_3Q fa18cbdd-bdf7-4a1e-a17b-84c89162612b          LGM45h
#> 19 LGM45h_4Q ca5454c4-b39f-4f04-bba9-419e081251e8          LGM45h
#> 20    LGM43h e7806cfa-c741-4ba6-98c7-631bc912d0c7            <NA>
#> 21 LGM43h_1Q 6f79288f-bcc1-4927-a456-acf043d9da08          LGM43h
#> 22 LGM43h_2Q 5245cc49-fead-4ac8-a095-4293fd112f5e          LGM43h
#> 23 LGM43h_3Q 66a161c6-5565-4914-901b-c636be2b40f8          LGM43h
#> 24 LGM43h_4Q cdf5e8b2-bc2b-4ffe-8688-2e5126908496          LGM43h
#> 25 LGM45h_1Q edc64488-5425-471b-bb19-824e1789a431          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot