Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r bb600d63-b27e-429d-b9c4-b7ca24c3cd75            <NA>
#> 2  LGM08r_1Q 1044ad46-ef16-4174-9bfb-f1149617adb3          LGM08r
#> 3  LGM08r_2Q af23834d-d69c-4b78-b1ac-9fc5c0986064          LGM08r
#> 4  LGM08r_3Q 83fad04f-d724-42ef-8382-5352b4fa1751          LGM08r
#> 5  LGM08r_4Q 4280c998-05d8-46f1-9587-3275626c0f45          LGM08r
#> 6     LGM16l 54399f14-6554-4e6c-887b-13db8c24491a            <NA>
#> 7  LGM16l_1Q 9f4fc967-7d19-4b38-98de-4a86f515970d          LGM16l
#> 8  LGM16l_2Q 12abb76d-39df-4950-90a2-3eca427977bb          LGM16l
#> 9  LGM16l_3Q e3f94961-e298-4f32-a4f0-0302b17faf47          LGM16l
#> 10 LGM16l_4Q 89a4140a-380c-426c-85c5-5cf47d3af1fe          LGM16l
#> 11    LGM38h 369775cd-e2a9-4950-b4af-674aed168062            <NA>
#> 12 LGM38h_1Q 88910c7c-3893-417e-9f29-5fc66cd31563          LGM38h
#> 13 LGM38h_2Q 517c376e-032c-49d8-877f-059105a250d8          LGM38h
#> 14 LGM38h_3Q 752e7a98-8123-4fc2-aeb1-b71a308eb001          LGM38h
#> 15 LGM38h_4Q 3a62da58-6dd0-42fa-8a84-c53141d8822c          LGM38h
#> 16    LGM45h 57d4f162-df02-4a79-ae0e-9948302aa3a9            <NA>
#> 17 LGM45h_2Q 53ef871c-9dbf-47d3-b59f-db81d37a4161          LGM45h
#> 18 LGM45h_3Q 78879d1d-78cd-4c1b-bea8-71bd4f7ecd11          LGM45h
#> 19 LGM45h_4Q 51df2850-41bd-4ca3-b4f6-0de7cf657224          LGM45h
#> 20    LGM43h b52575f5-ca63-4df3-a6cc-d69b6b6fadbd            <NA>
#> 21 LGM43h_1Q 69db206a-9440-46c6-a0f5-ed012603b200          LGM43h
#> 22 LGM43h_2Q 323cdc8f-4b6a-4a87-ad48-b7fcf6bdba0d          LGM43h
#> 23 LGM43h_3Q 2d22ac6d-9261-42e8-abf9-83c642b71f15          LGM43h
#> 24 LGM43h_4Q e0f3d64e-26dd-454f-8e6e-40c0eac264c9          LGM43h
#> 25 LGM45h_1Q c3efac09-eabc-467d-8d13-ac2eb2b6c2f4          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot