Generates unique identifiers

Generates unique identifiers for plot elements or plot observation elements

fillUniqueIdentifiers(target, element = "plot", replace.existing = FALSE)

Arguments

target

The initial object of class VegX to be modified

element

The Veg-X main element for which unique identifiers are needed: "plot" or "plotObservation".

replace.existing

A boolean flag to indicate whether newly generated identifiers should replace existing identifiers.

Value

The modified object of class VegX.

References

Wiser SK, Spencer N, De Caceres M, Kleikamp M, Boyle B & Peet RK (2011). Veg-X - an exchange standard for plot-based vegetation data

See also

Examples

# Load source data
data(mokihinui)

# Create a new Veg-X document with projects, plots and plot observations (no data)
mapping = list(projectTitle = "Project", plotName = "Plot", subPlotName = "Subplot",
               obsStartDate = "PlotObsStartDate", obsEndDate = "PlotObsStopDate")
x = addPlotObservations(newVegX(), moki_site, mapping = mapping)
#>  1 project(s) parsed, 1 new project(s) added.
#>  25 plot(s) parsed, 25 new plot(s) added.
#>  25 plot observation(s) parsed, 25 new plot observation(s) added.

# Examine the result
showElementTable(x, "plot")
#>     plotName relatedPlotName plotRelationship
#> 1     LGM08r            <NA>             <NA>
#> 2  LGM08r_1Q          LGM08r          subplot
#> 3  LGM08r_2Q          LGM08r          subplot
#> 4  LGM08r_3Q          LGM08r          subplot
#> 5  LGM08r_4Q          LGM08r          subplot
#> 6     LGM16l            <NA>             <NA>
#> 7  LGM16l_1Q          LGM16l          subplot
#> 8  LGM16l_2Q          LGM16l          subplot
#> 9  LGM16l_3Q          LGM16l          subplot
#> 10 LGM16l_4Q          LGM16l          subplot
#> 11    LGM38h            <NA>             <NA>
#> 12 LGM38h_1Q          LGM38h          subplot
#> 13 LGM38h_2Q          LGM38h          subplot
#> 14 LGM38h_3Q          LGM38h          subplot
#> 15 LGM38h_4Q          LGM38h          subplot
#> 16    LGM45h            <NA>             <NA>
#> 17 LGM45h_2Q          LGM45h          subplot
#> 18 LGM45h_3Q          LGM45h          subplot
#> 19 LGM45h_4Q          LGM45h          subplot
#> 20    LGM43h            <NA>             <NA>
#> 21 LGM43h_1Q          LGM43h          subplot
#> 22 LGM43h_2Q          LGM43h          subplot
#> 23 LGM43h_3Q          LGM43h          subplot
#> 24 LGM43h_4Q          LGM43h          subplot
#> 25 LGM45h_1Q          LGM45h          subplot

# Add unique identifiers
y = fillUniqueIdentifiers(x, "plot")

# Examine the result
showElementTable(y, "plot")
#>     plotName                 plotUniqueIdentifier relatedPlotName
#> 1     LGM08r b0d4b941-85ae-42ac-aa54-e07f38edf374            <NA>
#> 2  LGM08r_1Q 57798aef-a13b-4c3d-bea4-7e60507e40d1          LGM08r
#> 3  LGM08r_2Q 4b687ab2-53f0-4087-955f-59bb7737b09f          LGM08r
#> 4  LGM08r_3Q c213b4ba-d792-4f34-a01d-7e12d04be12d          LGM08r
#> 5  LGM08r_4Q 02fb08bd-c1fa-43c6-a25e-9b00851b1088          LGM08r
#> 6     LGM16l b1304ca2-a8bf-47cc-a661-e0008ab9d73a            <NA>
#> 7  LGM16l_1Q acaff61f-ad45-412a-9e63-8fa9dde7089e          LGM16l
#> 8  LGM16l_2Q 3c707646-2019-455d-b594-f5539356e971          LGM16l
#> 9  LGM16l_3Q cddf30de-7167-45a3-bce4-7a9c5f1c278a          LGM16l
#> 10 LGM16l_4Q f0556d47-eff3-4b73-8c12-b0701f398fcf          LGM16l
#> 11    LGM38h a3aa332d-4e24-443b-90a5-2373de30bd3e            <NA>
#> 12 LGM38h_1Q 69c859c3-da0e-40a7-a97c-024beabdd2a7          LGM38h
#> 13 LGM38h_2Q 97a38f5e-707f-48ab-8c39-56da27de23cc          LGM38h
#> 14 LGM38h_3Q a180bb7e-fdd9-43b3-a6d7-6e869fa116d4          LGM38h
#> 15 LGM38h_4Q ce5a9ab3-afb1-4631-bb5d-58dcb88588ac          LGM38h
#> 16    LGM45h dee6fae4-0254-496a-8a62-5babc3760d22            <NA>
#> 17 LGM45h_2Q 12e6e673-deae-4078-a3e8-3700ca54e88e          LGM45h
#> 18 LGM45h_3Q 4cc049f8-edfa-4f9d-b9e1-01cbefafadc9          LGM45h
#> 19 LGM45h_4Q 7372a8e8-d185-47d3-95d2-5e95ad6d6a0f          LGM45h
#> 20    LGM43h e6979e04-2cb6-4c29-bb14-3c241df22e4b            <NA>
#> 21 LGM43h_1Q d68e81db-80d2-4675-977d-17cbab9450e1          LGM43h
#> 22 LGM43h_2Q b348b7f6-5be5-4060-b1a2-3a612f9c76a6          LGM43h
#> 23 LGM43h_3Q 9100896c-d1cb-4bc3-a3ad-38589e4038ab          LGM43h
#> 24 LGM43h_4Q bef876a0-0fe5-4203-a727-06615705ae4c          LGM43h
#> 25 LGM45h_1Q eac70a44-19f9-45c2-b363-f6b1665817e0          LGM45h
#>    plotRelationship
#> 1              <NA>
#> 2           subplot
#> 3           subplot
#> 4           subplot
#> 5           subplot
#> 6              <NA>
#> 7           subplot
#> 8           subplot
#> 9           subplot
#> 10          subplot
#> 11             <NA>
#> 12          subplot
#> 13          subplot
#> 14          subplot
#> 15          subplot
#> 16             <NA>
#> 17          subplot
#> 18          subplot
#> 19          subplot
#> 20             <NA>
#> 21          subplot
#> 22          subplot
#> 23          subplot
#> 24          subplot
#> 25          subplot