Sets preferred taxon nomenclature

Sets (or resets) a preferred taxon nomenclature to the organism identities of a data set using a lookup table

Usage

setPreferredTaxonNomenclature(
  target,
  x,
  mapping,
  date.format = "%Y-%m-%d",
  missing.values = c(NA, ""),
  verbose = TRUE
)

Arguments

target

The initial object of class VegX to be modified

x

A data frame where each row corresponds to a different value of 'originalOrganismName'.

mapping

A named list whose elements are strings that correspond to column names in x. Names of the list should be:

originalOrganismName - A string with the original name given by the author of the data set (required).

preferredTaxonName - A string with the preferred taxon name forming the taxon (required).

interpretationSource - A string describing the source for the last nomenclature interpretation applied to this organism identity (i.e. the Plant List). (optional).

interpretationCitation - A string of the publication where nomenclature interpretation is explained (optional).

interpretationDate - Date of taxon nomenclature interpretation (see date.format) (optional).

interpretationParty - Name of the party that undertook nomenclature interpretation (optional).

date.format

A character string specifying the input format of dates (see as.Date).

missing.values

A character vector of values that should be considered as missing data (see details).

verbose

A boolean flag to indicate console output of the nomenclatural change process.

Value

The modified object of class VegX.

See also

Examples

# Load source data
data(mokihinui)

# Create new Veg-X document with aggregate organism observations
mapping = list(plotName = "Plot", obsStartDate = "PlotObsStartDate",
               taxonName = "NVSSpeciesName",
               stratumName = "Tier", cover = "Category")
coverscale = defineOrdinalScaleMethod(name = "Recce cover scale",
                   description = "Recce recording method by Hurst/Allen",
                   subject = "plant cover",
                   citation = "Hurst, JM and Allen, RB. (2007) 
                       The Recce method for describing New Zealand vegetation – Field protocols. 
                       Landcare Research, Lincoln.",
                   codes = c("P","1","2","3", "4", "5", "6"),
                   quantifiableCodes = c("1","2","3", "4", "5", "6"),
                   breaks = c(0, 1, 5, 25, 50, 75, 100),
                   midPoints = c(0.05, 0.5, 15, 37.5, 62.5, 87.5),
                   definitions = c("Presence", "<1%", "1-5%","6-25%", "26-50%", 
                                   "51-75%", "76-100%"))
strataDef = defineMixedStrata(name = "Recce strata",
                   description = "Standard Recce stratum definition",
                   citation = "Hurst, JM and Allen, RB. (2007) 
                     The Recce method for describing New Zealand vegetation – Field protocols. 
                     Landcare Research, Lincoln.",
                   heightStrataBreaks = c(0, 0.3,2.0,5, 12, 25, 50),
                   heightStrataNames = paste0("Tier ",1:6),
                   categoryStrataNames = "Tier 7",
                   categoryStrataDefinition = "Epiphytes")
x = addAggregateOrganismObservations(newVegX(), moki_tcv,
                        mapping = mapping,
                        methods = c(cover=coverscale),
                        stratumDefinition = strataDef)
#>  1 additional aggregate organism measurements found: Category.
#>  Measurement method 'Recce cover scale' added for 'cover'.
#>  Stratum definition method 'Recce strata' added.
#>  7 new stratum definitions added.
#>  5 plot(s) parsed, 5 new added.
#>  5 plot observation(s) parsed, 5 new added.
#>  148 organism names(s) parsed, 148 new added.
#>  148 organism identitie(s) parsed, 148 new added.
#>  33 stratum observation(s) parsed, 33 new added.
#>  582 record(s) parsed, 582 new aggregate organism observation(s) added.

# Inspect the original organism identities
head(showElementTable(x, "organismIdentity"))
#>             identityName   originalOrganismName taxon
#> 1  Dacrydium cupressinum  Dacrydium cupressinum  TRUE
#> 2    Weinmannia racemosa    Weinmannia racemosa  TRUE
#> 3       Myrsine salicina       Myrsine salicina  TRUE
#> 4        Cyathea smithii        Cyathea smithii  TRUE
#> 5 Prumnopitys ferruginea Prumnopitys ferruginea  TRUE
#> 6    Nothofagus truncata    Nothofagus truncata  TRUE

y = setPreferredTaxonNomenclature(x, moki_lookup,
           c(originalOrganismName = "NVSSpeciesName", preferredTaxonName = "PreferredSpeciesName"))
#>  Preferred taxon name was set on 148 organism identities.
#>  Preferred taxon name is now different than original organism name on 22 organism identities.
#>  21 new organism name(s) added.

# Inspect the modified organism identities
head(showElementTable(y, "organismIdentity"))
#>             identityName   originalOrganismName taxon     preferredTaxonName
#> 1  Dacrydium cupressinum  Dacrydium cupressinum  TRUE  Dacrydium cupressinum
#> 2    Weinmannia racemosa    Weinmannia racemosa  TRUE    Weinmannia racemosa
#> 3       Myrsine salicina       Myrsine salicina  TRUE       Myrsine salicina
#> 4        Cyathea smithii        Cyathea smithii  TRUE        Cyathea smithii
#> 5 Prumnopitys ferruginea Prumnopitys ferruginea  TRUE Prumnopitys ferruginea
#> 6    Fuscospora truncata    Nothofagus truncata  TRUE    Fuscospora truncata